backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.15.1
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator>=2.2.0
openai==1.75.0
httpx==0.28.1
openpyxl==3.1.5
pandas==2.2.3
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
